% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql_delete.R
\name{get_sql_delete}
\alias{get_sql_delete}
\title{transforms a data frame content into SQL delete commands}
\usage{
get_sql_delete(df, tablename)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble} as a source for SQL delete commands}

\item{tablename}{The name of the database table to delete from}
}
\description{
transforms a data frame content into SQL delete commands
}
