% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql_insert.R
\name{get_sql_insert}
\alias{get_sql_insert}
\title{transforms a data frame content into SQL insert commands}
\usage{
get_sql_insert(df, tablename, add_method = "copy")
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble} as a source for SQL update commands}

\item{tablename}{The name of the database table to update from}

\item{add_method}{Specifies the import method (copy (default) or insert)}
}
\description{
transforms a data frame content into SQL insert commands
}
