% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql_update.R
\name{get_sql_update}
\alias{get_sql_update}
\title{transforms a data frame content into SQL update commands}
\usage{
get_sql_update(df, key_col, tablename)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble} as a source for SQL update commands}

\item{key_col}{The primary key for replacements}

\item{tablename}{The name of the database table to update from}
}
\description{
transforms a data frame content into SQL update commands
}
