% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql_upsert.R
\name{get_sql_upsert}
\alias{get_sql_upsert}
\title{compares two data.frames and transforms the delta into SQL commands}
\usage{
get_sql_upsert(new_df, old_df, key_col, tablename, del_old = FALSE)
}
\arguments{
\item{new_df}{A \code{data.frame} or \code{tibble} specifiying the database table after the changes}

\item{old_df}{A \code{data.frame} or \code{tibble} with the current database table content}

\item{tablename}{The name of the database table to delete from}

\item{del_old}{Boolean flag to specify if missing data in new_df should be deleted}
}
\description{
compares two data.frames and transforms the delta into SQL commands
}
